requirements.txt
ruff 
black
tiktoken
openai
cohere
pypdf
langchain
python-dotenv
chromadb
weaviate-client
pandas